Force sensor integration
Here's one possibility of how to use one force sensor to measure force in both directions. 
The servo can pivot a tiny bit, and a small spring keeps a pre-load on the force sensor. 
Hopefully the force sensor is accurate and repeatable enough that we don't need an overly large dead band in the middle.